Step 1 of reducing Android essentials
This builds an app ("Barcode Parts") in the anythinglabs package.
It is -- brokenly -- portrait mode, but runs and won't step on any
of the Barcode Scanner bits (intents, etc.). The front light is
disabled. Localization, help, and all advanced features are
disabled. The idea is to continue reducing until we have something
that does the bare minimum of capturing a barcode and/or
displaying one.